% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbols.R
\name{ems}
\alias{ems}
\title{Wrap text in em dashes}
\usage{
ems(...)
}
\arguments{
\item{...}{Character. Text to go between em dashes,
and any other arguments to be passed to \code{\link[glue]{glue}}.}
}
\value{
Character. String surrounded in em dashes.

\code{ems} is a wrapper for \code{\link[glue]{glue}},
so you can use curly bracket notation and multiple character strings
just like with \code{\link[glue]{glue}}.
}
\description{
Wrap text in em dashes
}
\examples{
ems("hello world")
ems("hello ", "world")
fname <- "Darth"
lname <- "Vader"
ems("hello there {fname} {lname}")
ems(
  "hello {fname} {lname}, ",
  "I says to him"
)
}
