% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbols.R
\name{en}
\alias{en}
\title{Print an en-dash}
\usage{
en(from = "", to = "", format = NA, expand = F)
}
\arguments{
\item{from}{Character. If the en-dash separates two strings, the first one.}

\item{to}{Character. If the en-dash separates two strings, the second one.}

\item{format}{Function. If provided, formats \code{from} and \code{to}.}

\item{expand}{Logical. Add spaces?}
}
\value{
Character. The from and to strings, separated by an en-dash (UTF-8).

\code{en} is a wrapper for \code{\link[glue]{glue}},
so you can use curly bracket notation in \code{from} and \code{to},
just like with \code{\link[glue]{glue}}.
}
\description{
Print an en-dash
}
\examples{
library(lubridate)
library(scales)
en()
en(today(), today() \%m+\% months(1))
en(today(), today() \%m+\% months(1), format = date_format("\%B \%e, \%Y"))
en(1234, 5678)
en(1234, 5678, format = comma)
en(1234, 5678, format = comma, expand = TRUE)
}
