% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbols.R
\name{enye}
\alias{enye}
\title{Print an enye}
\usage{
enye(uppercase = FALSE)
}
\arguments{
\item{uppercase}{Logical. Use TRUE for uppercase, FALSE for lowercase.}
}
\value{
Character. A UTF-8 enye.
}
\description{
Print an enye
}
\examples{
enye()
enye(uppercase = TRUE)
}
