% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbols.R
\name{ish}
\alias{ish}
\title{Print a plus/minus character.}
\usage{
ish(margin = "", x = "", format = NA, margin_format = format,
  parentheses = FALSE, contract = FALSE)
}
\arguments{
\item{margin}{The number after the plus/minus. Can be a character.}

\item{x}{The number, if any, before the plus/minus. Can be a character.}

\item{format}{Function. If provided, formats \code{x}.}

\item{margin_format}{Function. If provided, formats \code{margin}.}

\item{parentheses}{Logical. Should \code{margin} be wrapped in parentheses?}

\item{contract}{Logical. If \code{x} is provided, should spaces be removed?}
}
\value{
Character. \code{x} plus or minus a \code{margin} (UTF-8).

\code{ish} uses \code{\link[glue]{glue}},
so you can use curly bracket notation in \code{x} and \code{margin},
just like with \code{\link[glue]{glue}}.
}
\description{
Print a plus/minus character.
}
\examples{
library(scales)
ish()
ish("5\%")
ish(.05, format = percent_format(accuracy = 1))
ish(
  x = 100,
  margin = .05,
  margin_format = percent_format(accuracy = 1),
  parentheses = TRUE
)
ish(
  x = 1e3,
  margin = .05,
  format = comma,
  margin_format = percent_format(accuracy = 1),
  parentheses = TRUE
)
ish(x = 5, margin = 5)
ish(x = 5, margin = 5, contract = TRUE)
}
