% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbols.R
\name{obelus}
\alias{obelus}
\title{Print an obelus}
\usage{
obelus(numerator = "", denominator = "", format = NA, contract = F)
}
\arguments{
\item{numerator}{Character. If the obelus separates two strings,
the first one.}

\item{denominator}{Character. If the obelus separates two strings,
the second one.}

\item{format}{Function. If provided,
formats \code{numerator} and \code{denominator}.}

\item{contract}{Logical. Remove spaces?}
}
\value{
Character. The numerator and denominator (if provided),
formatted if necessary, separated by an obelus (UTF-8).

\code{obelus} is a wrapper for \code{\link[glue]{glue}},
so you can use curly bracket notation
in \code{numerator} and \code{denominator},
just like with \code{\link[glue]{glue}}.
}
\description{
Print an obelus
}
\examples{
library(scales)
obelus()
obelus(1234, 5678)
obelus(1234, 5678, format = comma)
obelus(1234, 5678, format = comma, contract = TRUE)
}
