% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbols.R
\name{pilcrow}
\alias{pilcrow}
\title{Print a paragraph symbol}
\usage{
pilcrow(paragraph_number = "", nbsp = paragraph_number != "")
}
\arguments{
\item{paragraph_number}{Numeric or character value to print after the symbol.}

\item{nbsp}{Logical. Include a non-breaking space between the symbol
and the paragraph number? Defaults to TRUE if there is a paragraph number.}
}
\value{
Character (UTF-8): A pilcrow and, if indicated, a paragraph number.

For multiple paragraphs, two pilcrows are traditionally used.
Accordingly, the function returns two pilcrows if \code{paragraph_number}
includes an en-dash, a comma, or (not recommended) a hyphen,
or if \code{paragraph_number} is a vector of length greater than one.
In that last scenario, paragraphs will be separated by commas and spaces.
}
\description{
Print a paragraph symbol
}
\examples{
pilcrow()
pilcrow(8)
pilcrow("8")
pilcrow("8", nbsp = FALSE)
pilcrow(en("8", "9"))
pilcrow("8 - 9") # hyphen not recommended; use en-dash for ranges
pilcrow("8, 22")
pilcrow(c("8", "22"))
}
