% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbols.R
\name{quotes}
\alias{quotes}
\title{Wrap text in curly quotes}
\usage{
quotes(..., type = c("double", "single"), curly = TRUE)
}
\arguments{
\item{...}{Character. Text to go between curly quotes,
and any other arguments to be passed to \code{\link[glue]{glue}}.}

\item{type}{Character. Type of quotes: "single" or "double".}

\item{curly}{Logical. Set to FALSE for straight quotes.}
}
\value{
Character. String surrounded in quotation marks.

\code{quotes} is a wrapper for \code{\link[glue]{glue}},
so you can use curly bracket notation and multiple character strings
just like with \code{\link[glue]{glue}}.
}
\description{
Wrap text in curly quotes
}
\examples{
library(glue)
quotes("hello world")
quotes("hello world", type = "single")
quotes("hello ", "world", type = "single")
fname <- glue("I{enye()}igo")
lname <- "Montoya"
quotes("hello, my name is {fname} {lname}", type = "single")
quotes(
  "Hello, my name is {fname} {lname}. ",
  "You killed my father. Prepare to die."
)
quotes(
  "Hello, my name is {fname} {lname}. ",
  "I'm boring, and I use straight quotes.",
  curly = FALSE
)
}
