% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbols.R
\name{signs}
\alias{signs}
\title{Add plus signs or proper minus signs}
\usage{
signs(..., format = scales::number, plus_sign = FALSE)
}
\arguments{
\item{...}{Numeric vector and any other arguments passed to format.}

\item{format}{Function that formats a numeric vector,
such as \code{scales::number}, \code{scales::comma}, \code{scales::percent},
or a function returned by \code{scales::unit_format}
(all of which are documented at \code{\link[scales]{number_format}}).}

\item{plus_sign}{Logical. Should positive values start with plus signs?}
}
\value{
A \code{UTF-8} character vector
}
\description{
The true minus sign (UTF 2212) -- neither an em dash, nor an en dash,
nor the usual hyphen-minus -- is highly underrated.
It just makes everything look better!
This function builds on the formatting functions
from the \code{scales} package by replacing the hyphen-minus
with a true minus, and adding a plus sign if desired.
}
\examples{
signs(seq(-5, 5))
signs(seq(-5, 5), plus_sign = TRUE)
signs(seq(-5, 5) / 100, format = scales::percent, plus_sign = TRUE)
}
