% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbols.R
\name{silcrow}
\alias{silcrow}
\title{Print a section symbol}
\usage{
silcrow(section_number = "", nbsp = section_number != "")
}
\arguments{
\item{section_number}{Numeric or character value to print after the symbol.
Can also be a vector of such values.}

\item{nbsp}{Logical. Include a non-breaking space between the symbol
and the section number? Defaults to TRUE if there is a section number.}
}
\value{
Character (UTF-8): A silcrow and, if indicated, a section number.

For multiple sections, two silcrows are traditionally used.
Accordingly, the function returns two silcrows if \code{section_number}
includes an en-dash, a comma, or (not recommended) a hyphen,
or if \code{section_number} is a vector of length greater than one.
In that last scenario, sections will be separated by commas and spaces.
}
\description{
Print a section symbol
}
\examples{
silcrow()
silcrow(8.1)
silcrow("8.1.1")
silcrow("8.1.1", nbsp = FALSE)
silcrow(en("8.1", "8.2"))
silcrow("8.1 - 8.2") # hyphen not recommended; use en-dash for ranges
silcrow("8.1, 8.7")
silcrow(c("8.1", "8.7"))
}
