% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbols.R
\name{ticks}
\alias{ticks}
\title{Wrap text in backticks}
\usage{
ticks(..., r = FALSE)
}
\arguments{
\item{...}{Character. Text to go between backticks,
and any other arguments to be passed to \code{\link[glue]{glue}}.}

\item{r}{Logical. Add \code{r } at the beginning, for R Markdown?}
}
\value{
Character. String surrounded in backticks.

\code{ticks} is a wrapper for \code{\link[glue]{glue}},
so you can use curly bracket notation and multiple character strings
just like with \code{\link[glue]{glue}}.

\code{ticks} is particularly useful for R Markdown,
where you don"t want to use a backtick
because it would end the inline code block.

Sometimes you want to actually show an inline R code block, as text
(instead of evaluating it).
For that we"ve included a handy shortcut for adding the initial \code{r }.
}
\description{
Wrap text in backticks
}
\examples{
ticks("foobar")
ticks("foo", "bar")
word1 <- "foo"
word2 <- "bar"
ticks("{word1}{word2} = 'stupid'", r = TRUE)
word3 <- "random"
ticks(
  "c(",
  "{word1}{word2} = 'surprisingly', ",
  "{word3} = 'painless'",
  ")",
  r = TRUE
)
}
