% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbols.R
\name{times}
\alias{times}
\title{Print a multiplication symbol}
\usage{
times(multiplier = "", multiplicand = "", format = NA,
  contract = F)
}
\arguments{
\item{multiplier}{Character. If the multiplication symbol
separates two strings, the first one.}

\item{multiplicand}{Character. If the multiplication symbol
separates two strings, the second one.}

\item{format}{Function. If provided,
formats \code{multiplier} and \code{multiplicand}.}

\item{contract}{Logical. Remove spaces?}
}
\value{
Character. The multiplier and multiplicand (if provided),
formatted if necessary, separated by a multiplication symbol (UTF-8).

\code{times} is a wrapper for \code{\link[glue]{glue}},
so you can use curly bracket notation
in \code{multiplier} and \code{multiplicand},
just like with \code{\link[glue]{glue}}.
}
\description{
Print a multiplication symbol
}
\examples{
library(scales)
times()
times(1234, 5678)
times(1234, 5678, format = comma)
times(1234, 5678, format = comma, contract = TRUE)
}
